requirements.txt
fastapi
uvicorn
razorpay
python-dotenv
sqlalchemy
psycopg2-binary
reportlab
qrcode
pillow